CLJ-2003: Nesting cat inside ? causes unform to return nested result
The issue is actually any regex op under s/? is nested during unform.
Impl strategy is to check if nested op is a regex op if current op
is s/?; if so, do not wrap result in a vector. Nil is a special case
value when unforming s/? and returns an empty vector. Unfortunately
we cannot distinguish between a desired nil and no match:

    (is (nil? (s/conform (s/? nil?) [nil])))
    (is (nil? (s/conform (s/? nil?) [])))
    (is (nil? (s/conform (s/? nil?) nil)))
    (is (= (s/unform (s/? nil?) nil) []))

This patch does not address ::s/accept as first op under s/?: this
seems to be impossible. I am unsure if the ::s/accept case is
reachable at all in s/op-unform.